backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "SAP EWA Analyzer - AI-powered Early Watch Alert analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.74.0",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-storage-blob>=12.25.1",
    "docx2txt>=0.9",
    "fastapi>=0.115.12",
    "google-genai>=1.0.0",
    "jsonschema>=4.24.0",
    "openai>=1.82.0",
    "pymupdf4llm>=0.0.24",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
    "markdown2>=2.4.13",
    "WeasyPrint>=62.0",
    # SAP BTP Cloud Foundry dependencies
    "gunicorn>=21.0.0",
    "PyJWT>=2.8.0",
]
